cidra-server/requirements.txt
redis
python-dotenv
flask